Fix UI bugs and remove duplicate code
