--- src/components/Claim copy/steps/ClaimOverview/styles.module.css
@@ -1,14 +1,0 @@
-.maxButton {
-  border: 0;
-  background: none;
-  color: var(--mui-palette-primary-main);
-  padding: 0;
-  font-size: 16px;
-  font-weight: bold;
-  cursor: pointer;
-}
-
-.input :global .MuiInputBase-input {
-  padding: 12px 14px 12px 0;
-  font-weight: 700;
-}
