+++ src/components/Claim/index.tsx
@@ -33,6 +33,11 @@ import SafeToken from '@/public/images/token.svg'
 import css from './styles.module.css'
 import { useRouter } from 'next/router'
 import { AppRoutes } from '@/config/routes'
+import { canRedeemSep5Airdrop } from '@/utils/airdrop'
+import { Sep5InfoBox } from '../Sep5InfoBox'
+import { formatAmount } from '@/utils/formatters'
+import { ClaimCard } from '../ClaimCard'
+import { InfoAlert } from '../InfoAlert'
 
 const validateAmount = (amount: string, maxAmount: string) => {
   return mustBeFloat(amount) || minMaxValue(0, maxAmount, amount) || maxDecimals(amount, 18)
@@ -67,10 +72,17 @@ const ClaimOverview = (): ReactElement => {
   // Allocation, vesting and voting power
   const { data: allocation } = useSafeTokenAllocation()
 
-  const { investorVesting } = getVestingTypes(allocation?.vestingData ?? [])
+  const canRedeemSep5 = canRedeemSep5Airdrop(allocation)
 
-  const { sep5, user, investor, total } = useTaggedAllocations()
+  const { sep5Vesting, ecosystemVesting, investorVesting } = getVestingTypes(allocation?.vestingData ?? [])
+
+  const { sep5, user, ecosystem, investor, total } = useTaggedAllocations()
   const totalClaimableAmountInEth = formatEther(total.claimable)
+
+  const decimals = getDecimalLength(total.inVesting)
+
+  const hasAllocation = Number(total.allocation) > 0
+  const isClaimable = Number(total.claimable) > 0
 
   // Flags
   const isInvestorClaimingDisabled = !!investorVesting && isTokenPaused
@@ -124,83 +136,37 @@ const ClaimOverview = (): ReactElement => {
   return (
     <PaperContainer>
       <Grid container spacing={3} direction="row" justifyContent="space-evenly">
-        <Grid item xs={6}>
-          <Paper
-            sx={{
-              p: 3,
-              backgroundColor: ({ palette }) => palette.background.default,
-              color: ({ palette }) => palette.text.primary,
-              position: 'relative',
-              height: '100%',
-              display: 'flex',
-              flexDirection: 'column',
-              gap: 1,
-            }}
-          >
-            <Typography variant="subtitle1" fontWeight={700}>
-              Claim now
-            </Typography>
-            <Box display="inline-flex" gap={1} alignItems="center">
-              <Typography color="#B2BBC0" variant="body2" fontWeight={700}>
-                Total
-              </Typography>
-              <ShieldOutlinedIcon fontSize="small" />
-            </Box>
-            <TotalVotingPower />
-          </Paper>
-        </Grid>
-        <Grid item xs={6}>
-          <Paper
-            sx={{
-              p: 3,
-              backgroundColor: ({ palette }) => palette.background.default,
-              color: ({ palette }) => palette.text.primary,
-              position: 'relative',
-              height: '100%',
-              display: 'flex',
-              flexDirection: 'column',
-              gap: 1,
-            }}
-          >
-            <Typography variant="subtitle1" fontWeight={700}>
-              Claim at the end of the season
-              <Tooltip title={<Typography>tbd</Typography>} arrow placement="top">
-                <InfoOutlined
-                  sx={{
-                    height: '16px',
-                    width: '16px',
-                    mb: '-2px',
-                    ml: 1,
-                    color: 'var(--mui-palette-border-main)',
-                  }}
-                />
-              </Tooltip>
-            </Typography>
-            <Box display="inline-flex" gap={1} alignItems="center">
-              <Typography color="#B2BBC0" variant="body2" fontWeight={700}>
-                Total
-              </Typography>
-              <ShieldOutlinedIcon fontSize="small" />
-            </Box>
-            <TotalVotingPower />
-          </Paper>
+        <Grid item container xs={12} mb={1} spacing={3}>
+          <Grid item xs={12} sm={6}>
+            <ClaimCard
+              variant="claimable"
+              isGuardian={!!ecosystemVesting}
+              totalAmount={total.claimable}
+              ecosystemAmount={ecosystem.claimable}
+              decimals={decimals}
+            />
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <ClaimCard
+              variant="vesting"
+              isGuardian={!!ecosystemVesting}
+              totalAmount={total.inVesting}
+              ecosystemAmount={ecosystem.inVesting}
+              decimals={decimals}
+            />
+          </Grid>
         </Grid>
         <Grid item xs={12}>
           <Stack spacing={3}>
-            <Typography variant="body2" color="text.secondary">
-              <InfoOutlined
-                sx={{
-                  height: '16px',
-                  width: '16px',
-                  mb: '-2px',
-                  mr: 1,
-                }}
-              />
-              Total awarded allocation is{' '}
-              <Typography variant="body2" display="inline" color="text.primary">
-                10.000,15 SAFE
+            <InfoAlert>
+              <Typography variant="body2" color="text.secondary">
+                Total awarded allocation is{' '}
+                <Typography component="span" variant="inherit" color="text.primary">
+                  {formatAmount(formatEther(total.allocation), 2)} SAFE
+                </Typography>
               </Typography>
-            </Typography>
+            </InfoAlert>
+
             <Paper
               sx={{
                 p: 3,
@@ -223,7 +189,9 @@ const ClaimOverview = (): ReactElement => {
                 </Box>
               </Stack>
             </Paper>
+
             <Divider />
+
             <Box>
               <Typography variant="h6" fontWeight={700}>
                 How much do you want to claim?
@@ -232,33 +200,46 @@ const ClaimOverview = (): ReactElement => {
                 Select all tokens or define a custom amount.
               </Typography>
             </Box>
-            <Box display="inline-flex" gap={2}>
-              <TextField
-                variant="outlined"
-                value={amount}
-                error={!!amountError}
-                helperText={amountError}
-                onChange={onChangeAmount}
-                InputProps={{
-                  startAdornment: (
-                    <InputAdornment position="start" sx={{ width: '24px', height: '24px' }}>
-                      <SafeToken />
-                    </InputAdornment>
-                  ),
-                  endAdornment: (
-                    <InputAdornment position="end">
-                      <button onClick={setToMaxAmount} className={css.maxButton}>
-                        Max
-                      </button>
-                    </InputAdornment>
-                  ),
-                }}
-                className={css.input}
-              />
-              <Button variant="contained" onClick={onClaim} disableElevation disabled={isClaimDisabled}>
-                {creatingTxs ? <CircularProgress size={20} /> : 'Claim'}
-              </Button>
-            </Box>
+
+            {canRedeemSep5 && (
+              <Grid item xs={12}>
+                <Sep5InfoBox />
+              </Grid>
+            )}
+
+            <Grid item container gap={3} flexWrap="nowrap" xs={12} mb={1}>
+              <Grid item xs={4}>
+                <TextField
+                  variant="outlined"
+                  fullWidth
+                  value={amount}
+                  error={!!amountError}
+                  helperText={amountError}
+                  onChange={onChangeAmount}
+                  InputProps={{
+                    startAdornment: (
+                      <InputAdornment position="start" sx={{ width: '24px', height: '24px' }}>
+                        <SafeToken />
+                      </InputAdornment>
+                    ),
+                    endAdornment: (
+                      <InputAdornment position="end">
+                        <button onClick={setToMaxAmount} className={css.maxButton}>
+                          Max
+                        </button>
+                      </InputAdornment>
+                    ),
+                  }}
+                  className={css.input}
+                />
+              </Grid>
+
+              <Grid item xs={2}>
+                <Button variant="contained" fullWidth onClick={onClaim} disableElevation disabled={isClaimDisabled}>
+                  {creatingTxs ? <CircularProgress size={20} /> : 'Claim'}
+                </Button>
+              </Grid>
+            </Grid>
           </Stack>
         </Grid>
       </Grid>
