--- src/components/Intro copy/styles.module.css
@@ -1,8 +1,0 @@
-.overview {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-width: 182px;
-  height: 93px;
-}
